v0 terminé ouaiss

+++ DM2/src/waveform.c
@@ -4,7 +4,7 @@
 #include "sound.h"
 #include "waveform.h"
 
-sound_t* white(double T, int f_ech)
+sound_t* white(float T, int f_ech)
 {
     sound_t* s = malloc(sizeof(sound_t));
     s->n_samples = T * f_ech;
@@ -16,7 +16,7 @@ sound_t* white(double T, int f_ech)
     return s;
 }
 
-sound_t* sine(double f, int A, double T, int f_ech)
+sound_t* sine(float f, int A, float T, int f_ech)
 {
     sound_t* s = malloc(sizeof(sound_t));
     s->n_samples = T * f_ech;
@@ -28,20 +28,20 @@ sound_t* sine(double f, int A, double T, int f_ech)
     return s;
 }
 
-sound_t* square(double f, int A, double T, int f_ech)
+sound_t* square(float f, int A, float T, int f_ech)
 {
     sound_t* s = malloc(sizeof(sound_t));
     s->n_samples = T * f_ech;
     s->samples = malloc(s->n_samples * sizeof(int16_t));
     for (int i = 0; i < s->n_samples; ++i)
     {
-        double x = f * i / f_ech;
+        float x = f * i / f_ech;
         s->samples[i] = A * (2 * (floor(x) - floor(x - 0.5)) - 1);
     }
     return s;
 }
 
-sound_t* triangle(double f, int A, double T, int f_ech)
+sound_t* triangle(float f, int A, float T, int f_ech)
 {
     sound_t* s = malloc(sizeof(sound_t));
     s->n_samples = T * f_ech;
@@ -53,14 +53,14 @@ sound_t* triangle(double f, int A, double T, int f_ech)
     return s;
 }
 
-sound_t* sawtooth(double f, int A, double T, int f_ech)
+sound_t* sawtooth(float f, int A, float T, int f_ech)
 {
     sound_t* s = malloc(sizeof(sound_t));
     s->n_samples = T * f_ech;
     s->samples = malloc(s->n_samples * sizeof(int16_t));
     for (int i = 0; i < s->n_samples; ++i)
     {
-        double x = f * i / f_ech - 0.5;
+        float x = f * i / f_ech - 0.5;
         s->samples[i] = A * (2 * (x - floor(x)) - 1);
     }
     return s;
